Add go keyword syntax highlighting
This commit helps distinguish between keywords to allow for
better use of the theme. Specifically, the current scope of
"keyword" is too broad for a language like go as the control
flow keywords and various other keywords (func, struct, type,
interface, var, and const) blend together in Sunglo red.

In order to honor the original intent of the theme, the named
keywords above are colored in Cornflower Blue to help balance
the theme better. There was a distinct lack of Cornflower
and an over abundance of Sunglo prior to this change.

I have also confirmed that these changes do not impact the
javascript syntax highlighting per the author's guidelines.